Remove Attrs Call From getBucket
This change removes the Attrs call from the stroage client. The Attrs
call requires that the caller have storage.bucket.get permission,
typically given through roles/storage.bucketViewer.

Problem

Normally, this may be fine, but in a secured environment, the resource
owner may provide read-only access to a storage object and at the same
time may refuse to give access to the bucket attributes.

In this situation, the client is unable to use the gcsfs due to the
fact that "opening a file" reads bucket attributes.

Solution

The Attrs are not stored or used anywhere else. The call is used to
determine whether the user has access essentially to the bucket, which
as the problem statement explains, is not always the case when objects
are still readable.

Ultimately, the Attrs call is removed, and the other calls assume the
risk of not being able to get bucket data

Open Questions

There's a question of whether the existing functionality is good
practice, or trying to mimic the operation of a filesystem. However,
we need to be able to open files without needing bucket permissions.

This will accept the risk of any non-working buckets and pass that off
to the object manipulation functions